+++ lib/WorseReflection/Tests/Inference/TestAssertWalker.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\WorseReflection\Tests\SelfTest;
+namespace Phpactor\WorseReflection\Tests\Inference;
 
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\Expression\ArgumentExpression;
