+++ lib/WorseReflection/Tests/Inference/function/is_int.test
