+++ lib/WorseReflection/Tests/Inference/function/is_string.test
