+++ lib/WorseReflection/Tests/Inference/generics/array_access1.test
