+++ lib/WorseReflection/Tests/Inference/generics/array_access_resolve_method_type1.test
