+++ lib/WorseReflection/Tests/Inference/generics/class_extend1.test
