+++ lib/WorseReflection/Tests/Inference/generics/class_extend2.test
