+++ lib/WorseReflection/Tests/Inference/generics/class_implements_multiple1.test
