+++ lib/WorseReflection/Tests/Inference/generics/class_implements_single1.test
