+++ lib/WorseReflection/Tests/Inference/generics/class_template_extends1.test
