+++ lib/WorseReflection/Tests/Inference/generics/class_template_implements1.test
