+++ lib/WorseReflection/Tests/Inference/generics/iterator1.test
