+++ lib/WorseReflection/Tests/Inference/generics/iterator2.test
