+++ lib/WorseReflection/Tests/Inference/generics/iterator_aggregate1.test
