+++ lib/WorseReflection/Tests/Inference/generics/iterator_aggregate2.test
