+++ lib/WorseReflection/Tests/Inference/generics/method_returns_collection.test
