+++ lib/WorseReflection/Tests/Inference/generics/method_returns_collection2.test
