+++ lib/WorseReflection/Tests/Inference/if-statement/die.test
