+++ lib/WorseReflection/Tests/Inference/if-statement/false.test
@@ -1,0 +1,8 @@
+<?php
+
+if (false === $foobar instanceof Foobar) {
+    wrAssertType('<missing>', $foobar);
+    die();
+}
+
+wrAssertType('Foobar', $foobar);
