+++ lib/WorseReflection/Tests/Inference/if-statement/instanceof.test
@@ -1,0 +1,6 @@
+<?php
+
+if ($foobar instanceof Foobar) {
+    wrAssertType('Foobar', $foobar);
+}
+wrAssertType('<missing>', $foobar);
