+++ lib/WorseReflection/Tests/Inference/if-statement/multiple_statements_with_class.test
