+++ lib/WorseReflection/Tests/Inference/if-statement/non-terminating-branch.test
@@ -8,4 +8,4 @@ if (is_int($f)) {
     wrAssertType('int', $f);
 }
 
-wrAssertType('int', $f);
+wrAssertType('<missing>', $f);
