+++ lib/WorseReflection/Tests/Inference/if-statement/type_after_break.test
@@ -1,0 +1,10 @@
+<?php
+
+
+foreach ([1, 2] as $hello) {
+    if (!$foobar instanceof Foobar) {
+        break;
+    }
+
+    wrAssertType('Foobar', $foobar);
+}
