+++ lib/WorseReflection/Tests/Inference/if-statement/type_after_exception.test
@@ -1,0 +1,7 @@
+<?php
+
+if ($foobar instanceof Foobar) {
+    throw new Exception("HAI!");
+}
+wrAssertType('<missing>', $foobar);
+
