+++ lib/WorseReflection/Tests/Inference/if-statement/type_after_return.test
@@ -1,0 +1,13 @@
+<?php
+
+function foo(): Foo|Bar {}
+
+$foobar = foo();
+
+wrAssertType('Foo|Bar', $foobar);
+
+if ($foobar instanceof Foo) {
+    return;
+}
+
+wrAssertType('Bar', $foobar);
